[refactor]: 옷장 카드 UI 폴더 이동
